% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_distance.R
\name{axis_distance}
\alias{axis_distance}
\title{Compute the largest distance between pixels with FALSE value}
\usage{
axis_distance(bi_array)
}
\arguments{
\item{bi_array}{an array selected from a binary matrix}
}
\value{
the largest distance between pixels with FALSE value on bi_array
}
\description{
This function return the largest distance between pixels with FALSE value in an array selected from a binary matrix.
}
