% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution.R
\name{convolution}
\alias{convolution}
\title{Compute the two-dimensional convolution}
\usage{
convolution(input_data, use_kernel, padding_value = 0, stride = 1)
}
\arguments{
\item{input_data}{a two-dimension matrix that need to perform convolution method on}

\item{use_kernel}{the kernel that used to compute the two-dimension convolution}

\item{padding_value}{how many zero-value rows and columns are supposed to be added before computing, with default = 0}

\item{stride}{the stride size to take when computing, with default = 1}
}
\value{
res the convolution result
}
\description{
This function return the result of two-dimensional convolution on the input data with input kernel
}
