% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_label.R
\name{detect_label}
\alias{detect_label}
\title{Detect the position of axis in binary image}
\usage{
detect_label(bi_matrix)
}
\arguments{
\item{bi_matrix}{a two-dimension binary matrix generated by a binary image}
}
\value{
y_position the column index of the y-axis position

x_position the row index of the x-axis position
}
\description{
This function return the result of the position of y-axis and x-axis separately in a given binary image
}
