% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_binary_matrix.R
\name{get_binary_matrix}
\alias{get_binary_matrix}
\title{Generate a binary matrix from a multi-channel pixel matrix.}
\usage{
get_binary_matrix(pixel_matrix, threshold = 0.8)
}
\arguments{
\item{pixel_matrix}{a three-dimension  matrix generated by a multiple channels image}

\item{threshold}{a number between 0 and 1. For pixel with value greater that threshold will be TRUE in return}
}
\value{
binary_matrix a binary matrix with the same dimension as the input matrix
}
\description{
This function generate a binary_matrix from pixel_matrix with multiple channels like RGB or RGBA.
}
