% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_label_position.R
\name{get_label_position}
\alias{get_label_position}
\title{Get the position of extreme value on Axis}
\usage{
get_label_position(binary_matrix)
}
\arguments{
\item{binary_matrix}{a two-dimension binary matrix generated by a binary image}
}
\value{
ymax_position the row index of the maximum y axis value on binary pixel image

ymin_position the row index of the minimum y axis value on binary pixel image

xmax_position the column index of the maximum x axis value on binary pixel image

xmin_position the column index of the minimum x axis value on binary pixel image
}
\description{
This function returns the position of extreme value on Axis
}
