% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_point_position.R
\name{get_point_position}
\alias{get_point_position}
\title{Get the position of data points on binary image}
\usage{
get_point_position(binary_matrix)
}
\arguments{
\item{binary_matrix}{a two-dimension binary matrix generated by a binary image}
}
\value{
data_position a two-column matrix with the position of the data points.
}
\description{
This function returns the position of data points on Axis by using convolution method.
}
