% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix2plot.R
\name{matrix2plot}
\alias{matrix2plot}
\title{Plot the image of a given matrix}
\usage{
matrix2plot(pixel_matrix)
}
\arguments{
\item{pixel_matrix}{a three-dimension matrix that can generate a multiple channels image}
}
\description{
This function plots the image of a given matrix as a multiple channel image.
}
