% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padding.R
\name{padding}
\alias{padding}
\title{Add zero-value rows and columns to each edges of matrix}
\usage{
padding(input_data, padding_value)
}
\arguments{
\item{input_data}{a two-dimension matrix that need to perform padding method on}

\item{padding_value}{how many zero-value rows and columns are supposed to be added}
}
\value{
res the padding result
}
\description{
This function return matrix with additional zero-value rows and columns on each edge
}
