% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position2value.R
\name{position2value}
\alias{position2value}
\title{Get the value of data}
\usage{
position2value(data_position, label_position, ymax, ymin, xmin, xmax)
}
\arguments{
\item{data_position}{a two-column matrix with the position of the data points.}

\item{label_position}{a list with length=4, with the position of the extreme value on corresponding axis.}

\item{ymax}{the maximum y value showing on the plot}

\item{ymin}{the minimum y value showing on the plot}

\item{xmin}{the maximum x value showing on the plot}

\item{xmax}{the minimum x value showing on the plot}
}
\value{
data a two-column matrix with the predicted value of the data points.
}
\description{
This function returns the value of detected data set based on the position and axis extreme value information
}
