% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter2data.R
\name{scatter2data}
\alias{scatter2data}
\title{Detect the value of a scatter plot}
\usage{
scatter2data(scatter_file, ymax, ymin, xmax, xmin)
}
\arguments{
\item{scatter_file}{the address of the scatter plot file.}

\item{ymax}{the maximum y value showing on the plot}

\item{ymin}{the minimum y value showing on the plot}

\item{xmax}{the minimum x value showing on the plot}

\item{xmin}{the maximum x value showing on the plot}
}
\value{
a two-column matrix with the predicted value of the data points.
}
\description{
This function returns the value of detected data set based on a standard scatter plot generated by ggplot::geom_point function.
}
