% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_one_channel.R
\name{to_one_channel}
\alias{to_one_channel}
\title{select one channel from a multiple channel matrix}
\usage{
to_one_channel(pixel_matrix)
}
\arguments{
\item{pixel_matrix}{a three-dimension matrix generated by a multiple channels image}
}
\value{
a two-dimension matrix
}
\description{
This function returns the value of detected data set based on the position and axis extreme value information
}
